Added swerve drive autonomous and teleop actions and added action deque test
Added one autonomous swerve drive action thread that follows a simple trajectory and added one teleop swerve drive action thread that will be the default swerve drive action thread. Added an action deque test that tests the canceling of action threads that will loop, which led to a fix for the action deque that fixed the action deque pausing the main thread forever.